// frontend/src/shared/api/generated/models/service-auth-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RIP Go API
 * API for cargo transportation service
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DsUser } from './ds-user';

export interface ServiceAuthResponse {
    'access_token'?: string;
    'expires_at'?: string;
    'refresh_token'?: string;
    'user'?: DsUser;
}