// frontend/src/shared/api/generated/models/service-login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RIP Go API
 * API for cargo transportation service
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ServiceLoginRequest {
    'login': string;
    'password': string;
}