// frontend/src/shared/api/generated/models/service-register-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RIP Go API
 * API for cargo transportation service
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ServiceRegisterRequest {
    'email': string;
    'login': string;
    'name': string;
    'password': string;
    'phone'?: string;
    'role'?: string;
}